backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cs-risk-agent"
version = "0.1.0"
description = "Enterprise Multi-Cloud AI Orchestrator for Consolidated Subsidiary Risk Analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "GRC Advisory Team" },
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",

    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # AI / LLM Providers
    "langchain>=0.3.10",
    "langgraph>=0.2.0",
    "azure-ai-inference>=1.0.0b7",
    "boto3>=1.35.0",
    "google-cloud-aiplatform>=1.72.0",
    "ollama>=0.4.0",

    # ML & Analysis
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "statsmodels>=0.14.0",
    "shap>=0.46.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "optuna>=4.1.0",

    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",

    # Observability
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "prometheus-client>=0.21.0",

    # Reports
    "reportlab>=4.2.0",
    "python-pptx>=1.0.0",
    "jinja2>=3.1.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # ETL / Data
    "httpx>=0.28.0",
    "lxml>=5.3.0",
    "openpyxl>=3.1.0",

    # Config
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "factory-boy>=3.3.0",
    "faker>=33.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cs_risk_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "T20", # flake8-print
    "RUF", # ruff-specific
]
ignore = ["S101", "S105", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["cs_risk_agent"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/cs_risk_agent --cov-report=term-missing --cov-report=html"
